aws: terraform provider implementation
Define a Provider that implements the terraform.Provider interface and
provide a NewTerraformProvider func that matches the signature of the
terraform.ProviderInitializer type. This allows it to be used as one of
the providers during plan reading process.

The Provider returns a query.Component slice for each terraform.Resource
that uses the Provider.