aws: add fully qualified name and update the use of Provider func
In Terraform 0.13+ the "aws" provider name changed to fully qualified
"registry.terraform.io/hashicorp/aws". The fix for this is to add the
RegistryName with this value to the list of MatchNames.

Also, due to the terraform.ProviderInitializer struct changing, the
Provider function needs to be updated to accept a map of configuration
values instead of the terraform.ProviderConfig struct.