Fix problems with Terraform 0.13+ and usage of variables (#20)
* terraform: implement fixes for #18 and #19

Add fields to Terraform plan schema to handle variable references in
provider configuration. Use the new `evaluateProviderConfigExpressions`
to retrieve expression values, instead of depending on ConstantValue,
fixing the problem in #19.

Additionally, place the initialized provider under each key from the
ProviderInitializer's MatchNames slice, enabling a fix for #18.

* aws: add fully qualified name and update the use of Provider func

In Terraform 0.13+ the "aws" provider name changed to fully qualified
"registry.terraform.io/hashicorp/aws". The fix for this is to add the
RegistryName with this value to the list of MatchNames.

Also, due to the terraform.ProviderInitializer struct changing, the
Provider function needs to be updated to accept a map of configuration
values instead of the terraform.ProviderConfig struct.

* e2e: fix broken tests