cost: continue if an error is encountered in NewState
Instead of returning the error and failing the entire process, NewState
will now store the error into the Component's `Error` field and
continue with the rest of the resources.

Additional methods on the ComponentDiff (Valid) and ResourceDiff
(Valid and Errors) can be used to check whether there are any errors
with a specific Component or one of the Components of a Resource.