testdata: reorganize test-data and add SQL dump
The testdata has been re-arranged to better reflect which providers where
being tested; this allows to deal more easily with tests and the
providers which are not yet supported.

Some SQL dump of a previous ingestion has also been added to be able to
test things locally more easily, it will be injected into the database
whenever this one is started. The database can now be access via `make
db-cli`