terraform: fix unexpected error with multiple providers
If using another provider (e.g. 'random'), the execution would stop
because of the recursion of the function, so using the flag logic
doesn't work as expected when multiple providers are mixed.

Instead that logic has been extracted afterwards, to ensure that at
least one of the queries is of a known provider's type.