aws/ingester: use variable name as csv column when ingest
To keep consistency between CSV value and query from resource
implementation. Ingester will keep the same value from the CSV to
the DB.
it impact previous key/column names and on_demand/reserved value which actually is OnDemand and Reserved
from the CSV.

Column name will be change with field variable name when ingesting in
the DB.
Basically it is the name without spacies.
Even if json accept space in key names, we choose to remove them for
simplification reason.